[FiX/VG-28] 캐스팅 수정

+++ Source/GA6thFinal_Framework/GameScripts/Scripts/TurnSystem/TurnActor/Character/Player/Condition/PlayerDeadCondition.cpp
@@ -19,7 +19,7 @@ void PlayerDeadCondition::OnStart()
 
 bool PlayerDeadCondition::Evaluate()
 {
-    float hp = GetPlayer().HP;
+    int hp = GetPlayer().HP;
     bool isDead = 0 >= hp;
     return isDead;
 }
